+++ frontend/src/router/index.ts
@@ -5,7 +5,7 @@ import SearchView from '../views/SearchView.vue'
 import StatisticsView from '../views/StatisticsView.vue'
 import SettingsView from '../views/SettingsView.vue'
 import TagManagerView from '../views/TagManagerView.vue'
-import GameWallView from '../views/GameWallView.vue'
+import GameWallView from '../views/game/GameWallView.vue'
 
 const routes: RouteRecordRaw[] = [
   { path: '/', name: 'home', component: HomeView },